refactor(@schematics/angular): remove private `/components` API
Last usages of this API has been removed in https://github.com/angular/components/pull/28424
