+++ packages/schematics/angular/package.json
@@ -13,8 +13,7 @@
     "./utility": "./utility/index.js",
     "./utility/*": "./utility/*.js",
     "./migrations/migration-collection.json": "./migrations/migration-collection.json",
-    "./*": "./*.js",
-    "./private/components": "./private/components.js"
+    "./*": "./*.js"
   },
   "schematics": "./collection.json",
   "dependencies": {
