--- packages/schematics/angular/private/components.ts
@@ -1,15 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-export {
-  addModuleImportToStandaloneBootstrap,
-  addFunctionalProvidersToStandaloneBootstrap,
-  callsProvidersFunction,
-  findBootstrapApplicationCall,
-  importsProvidersFrom,
-} from './standalone';
